rename project-type -> types.ts
