+++ src/project/types/book/book-render.ts
@@ -43,7 +43,7 @@ import { outputRecipe } from "../../../command/render/output.ts";
 import { renderCleanup } from "../../../command/render/cleanup.ts";
 
 import { ProjectConfig, ProjectContext } from "../../types.ts";
-import { ProjectOutputFile } from "../project-type.ts";
+import { ProjectOutputFile } from "../types.ts";
 
 import { executionEngineKeepMd } from "../../../execute/engine.ts";
